Definicion de los vertices del cuadrado
Algunos cambios en la clase vertice (float)

Metodo auxiliar en la clase prisma- distancia entre dos puntos.

Un main para probar los metodos

+++ ProyectoFinalP1/src/logico/Vertice.java
@@ -2,28 +2,28 @@ package logico;
 
 public class Vertice {
 	
-	private int x;
-	private int y;
+	private float x;
+	private float y;
 	
-	public Vertice(int x, int y) {
+	public Vertice(float x, float y) {
 		super();
 		this.x = x;
 		this.y = y;
 	}
 
-	public int getX() {
+	public float getX() {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
